+++ cooked_input/tests/test_get_int.py
@@ -16,7 +16,13 @@ pytest stuff
 Len Wanger, 2017
 """
 
-from io import StringIO
+import sys
+
+if sys.version_info[0] > 2:  # For Python 3
+    from io import StringIO
+else:
+    from StringIO import StringIO
+
 from cooked_input import get_input, get_int, silent_error, log_error
 from cooked_input import IntConvertor, RangeValidator, EqualToValidator
 from cooked_input import NoneOfValidator, AnyOfValidator