+++ cooked_input/tests/test_get_str.py
@@ -16,7 +16,13 @@ pytest stuff
 Len Wanger, 2017
 """
 
-from io import StringIO
+import sys
+
+if sys.version_info[0] > 2:  # For Python 3
+    from io import StringIO
+else:
+    from StringIO import StringIO
+
 from .utils import redirect_stdin
 
 from cooked_input import get_input