+++ pytest_coverage.sh
@@ -1,2 +1,4 @@
 
-pytest --cov=cooked_input cooked_input --cov=cooked_input cooked_input/
+
+# needs pytest-cov to be pip installed
+pytest --cov-append --cov=cooked_input cooked_input --cov=cooked_input cooked_input/