+++ pytest_coverage_report.sh
@@ -1,3 +1,3 @@
 
 
-pytest --cov-report term-missing:skip-covered --cov=cooked_input cooked_input/
+pytest --cov-append --cov-report term-missing:skip-covered --cov=cooked_input cooked_input/